chore: move AlreadyRedeemedError and StaffPassNotFoundError classes to errors folder
