+++ src/errors/alreadyRedeemedError.ts
@@ -1,0 +1,6 @@
+export class AlreadyRedeemedError extends Error {
+  constructor(teamName: string) {
+    super(`Team has already redeemed: ${teamName}`);
+    this.name = 'AlreadyRedeemedError';
+  }
+}
