+++ src/errors/staffPassNotFoundError.ts
@@ -1,0 +1,6 @@
+export class StaffPassNotFoundError extends Error {
+  constructor(staffPassId: string) {
+    super(`Staff pass ID not found: ${staffPassId}`);
+    this.name = 'StaffPassNotFoundError';
+  }
+}
